bot.errbot-slack/requirements.txt
ansi==0.1.3
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bottle==0.12.13
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
colorama==0.3.9
colorlog==3.1.2
cryptography==2.0.3
dnspython==1.15.0
dnspython3==1.15.0
errbot==5.1.3
gitdb2==2.0.3
GitPython==2.1.9
idna==2.6
invoke==0.22.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
pathlib==1.0.1
pycparser==2.18
Pygments==2.2.0
pygments-markdown-lexer==0.1.0.dev39
pyOpenSSL==17.5.0
requests==2.18.4
rocket-errbot==1.2.5
six==1.11.0
slackclient==1.2.1
smmap2==2.0.3
tailer==0.4.1
urllib3==1.22
virtualenv==15.2.0
waitress==1.1.0
WebOb==1.7.4
websocket-client==0.47.0
WebTest==2.0.29
Yapsy==1.11.223
